==> notebooks/03_Shiny_Experiments.Rmd <==
---
title: "R Notebook"
output: html_notebook
author: Marco Polo Bravo Montiel
date: 2020-04-23
---

# Shiny App Experiment Lab

### Libraries

```{r}
library(here)
library(dplyr)
library(arrow)
library(tidyr)
library(geojsonio)
```

### Reading data

```{r}
census_dataset <- open_dataset(here("data", "processed", "parquet_data_coords"))
census_dataset

```

### Reading specific data

```{r}
pueb_norm <- census_dataset |>
    filter(NOM_ENT=="Puebla") |> 
    collect()

pueb_norm
```

```{r}
unique(pueb_norm$NOM_MUN)
```

### Example

```{r}
extract_coordinates <- function(data, municipality, locality) {
  selected_location <- data |> 
    filter(NOM_MUN == municipality, NOM_LOC == locality)
  
  coordinates <- tibble(
    long = selected_location$longitude_decimal,
    lat = selected_location$latitude_decimal
  )
  
  return(coordinates)
}


municipality <- "Acajete"
locality <- "Santa Isabel Tepetzala"

red_point <- extract_coordinates(pueb_norm, municipality, locality) |> 
  slice(1)

# red_point <- data.frame(long = -98.2035, lat = 19.0414)
geojson_file <- geojson_read("../data/processed/mexico.geojson",  what = "sp")

filtered_geojson <- geojson_file |> 
  filter(name == "Puebla")

ggplot() +
  geom_polygon(data = filtered_geojson,
               aes(x = long, y = lat, group = group),
               fill = "lightgray", color = "white") +
  geom_point(data = red_point, aes(x = long, y = lat), color = "red", size = 3) +
  theme_void() +
  coord_map()
```

### Population graph

```{r}
pueb_norm
```

### Getting column names

```{r}
column_names <- names(pueb_norm)
column_names
```

### Getting column names that have to do with population age cohorts (Masculine/Feminine)

```{r}
matching_columns <- grep("^P_.*[MF]$", column_names, value = TRUE)
matching_columns
```

### Separate by sex

```{r}
ending_in_M <- character(0)
ending_in_F <- character(0)

for (col_name in matching_columns) {
  if (endsWith(col_name, "M")) {
    ending_in_M <- c(ending_in_M, col_name)
  } else if (endsWith(col_name, "F")) {
    ending_in_F <- c(ending_in_F, col_name)
  }
}

print("Column names ending in M:")
print(ending_in_M)

print("Column names ending in F:")
print(ending_in_F)
```

```{r}
cohort_names_m <- c("P_0A4_M",
                    "P_5A9_M",
                    "P_10A14_M",
                    "P_15A19_M",
                    "P_20A24_M",
                    "P_25A29_M",
                    "P_30A34_M",
                    "P_35A39_M",
                    "P_40A44_M",
                    "P_45A49_M",
                    "P_50A54_M",
                    "P_55A59_M",
                    "P_60A64_M",
                    "P_65A69_M",
                    "P_70A74_M",
                    "P_75A79_M",
                    "P_80A84_M",
                    "P_85YMAS_M")

cohort_names_f <- c("P_0A4_F",
                    "P_5A9_F",
                    "P_10A14_F",
                    "P_15A19_F",
                    "P_20A24_F",
                    "P_25A29_F",
                    "P_30A34_F",
                    "P_35A39_F",
                    "P_40A44_F",
                    "P_45A49_F",
                    "P_50A54_F",
                    "P_55A59_F",
                    "P_60A64_F",
                    "P_65A69_F", 
                    "P_70A74_F",
                    "P_75A79_F",
                    "P_80A84_F", 
                    "P_85YMAS_F")



municipality <- "Acajete"
locality <- "San Javier"

pueb_norm_filt <- pueb_norm |> 
    filter(NOM_MUN == municipality, NOM_LOC == locality)

cohort_counts_m <- as.numeric(pueb_norm_filt[1,cohort_names_m])
cohort_counts_f <- as.numeric(pueb_norm_filt[1,cohort_names_f])

data <- tibble(
  Cohort = c(cohort_names_m, cohort_names_f),
  Count = c(cohort_counts_m, cohort_counts_f),
  Sex = rep(c("Male", "Female"), each = length(cohort_names_m))
)

# Plotting population pyramid
ggplot(data, aes(x = reorder(Cohort, -Count), y = Count, fill = Sex)) +
  geom_bar(stat = "identity", position = "identity") +
  scale_fill_manual(values = c("blue", "pink")) +  
  coord_flip() +  
  labs(title = "Population Pyramid",
       x = "Population Count",
       y = "Age Cohort",
       fill = "Sex") +
  theme_minimal()  

```

```{r}
new_ages <- c("0-4",
                 "5-9",
                 "10-14",
                 "15-19",
                 "20-24",
                 "25-29", 
                 "30-34", 
                 "35-39", 
                 "40-44", 
                 "45-49",
                 "50-54",
                 "55-59",
                 "60-64", 
                 "65-69",
                 "70-74", 
                 "75-79", 
                 "80-84", 
                 "85+")

data <- tibble(
  Age = paste0(new_ages),
  Male = sample(200:1000, length(cohort_names_m), replace = TRUE),
  Female = sample(200:1000, length(cohort_names_f), replace = TRUE)
)

data_long <- pivot_longer(
  data, 
  cols = c(Male, Female), 
  names_to = "Sex", 
  values_to = "Population"
)

basic_plot <- ggplot(data_long, aes(x = Age, y = ifelse(Sex == "Male", -Population, Population), fill = Sex)) +
  geom_bar(stat = "identity") +
  scale_y_continuous(labels = abs, limits = max(data_long$Population) * c(-1, 1)) +
  coord_flip() +
  theme_minimal() +
  labs(x = "Age", y = "Population", fill = "Sex", title = "Population Pyramid")

basic_plot
```

### Complete pipeline

```{r}
census_dataset <- open_dataset(here("data", "processed", "parquet_data_coords"))


pueb_norm <- census_dataset |>
    filter(NOM_ENT=="Puebla") |> 
    collect()

municipality <- "Acajete"
locality <- "San Javier"

pueb_norm_filt <- pueb_norm |> 
    filter(NOM_MUN == municipality, NOM_LOC == locality)


cohort_names_m <- c("P_0A4_M",
                    "P_5A9_M",
                    "P_10A14_M",
                    "P_15A19_M",
                    "P_20A24_M",
                    "P_25A29_M",
                    "P_30A34_M",
                    "P_35A39_M",
                    "P_40A44_M",
                    "P_45A49_M",
                    "P_50A54_M",
                    "P_55A59_M",
                    "P_60A64_M",
                    "P_65A69_M",
                    "P_70A74_M",
                    "P_75A79_M",
                    "P_80A84_M",
                    "P_85YMAS_M")

cohort_names_f <- c("P_0A4_F",
                    "P_5A9_F",
                    "P_10A14_F",
                    "P_15A19_F",
                    "P_20A24_F",
                    "P_25A29_F",
                    "P_30A34_F",
                    "P_35A39_F",
                    "P_40A44_F",
                    "P_45A49_F",
                    "P_50A54_F",
                    "P_55A59_F",
                    "P_60A64_F",
                    "P_65A69_F", 
                    "P_70A74_F",
                    "P_75A79_F",
                    "P_80A84_F", 
                    "P_85YMAS_F")



new_ages <- c("0-4",
                 "5-9",
                 "10-14",
                 "15-19",
                 "20-24",
                 "25-29", 
                 "30-34", 
                 "35-39", 
                 "40-44", 
                 "45-49",
                 "50-54",
                 "55-59",
                 "60-64", 
                 "65-69",
                 "70-74", 
                 "75-79", 
                 "80-84", 
                 "85+")


data <- tibble(
  Age = paste0(new_ages),
  Male = as.numeric(pueb_norm_filt[1,cohort_names_m]),
  Female = as.numeric(pueb_norm_filt[1,cohort_names_f])
)

data_long <- pivot_longer(
  data, 
  cols = c(Male, Female), 
  names_to = "Sex", 
  values_to = "Population"
)
basic_plot <- ggplot(data_long, aes(x = Age, y = ifelse(Sex == "Male", -Population, Population), fill = Sex)) +
  geom_bar(stat = "identity") +
  scale_y_continuous(labels = abs, limits = max(data_long$Population) * c(-1, 1)) +
  coord_flip() +
  theme_minimal() +
  labs(x = "Age", y = "Population", fill = "Sex", title = "Population Pyramid")

basic_plot
```

### Card

```{r}
card <- census_dataset |>
    filter(
      NOM_ENT == "Puebla",
      NOM_MUN == "Acateno",
      NOM_LOC == "Santa Andrea"
    ) |>
  collect()
    
    total_population <- as.numeric(card[1, c("POBTOT")], na.rm = TRUE)
    paste("Total Population:", total_population)
```

### Pie

```{r}
tot <- census_dataset |>
    filter(NOM_ENT=="Total nacional") |> 
    collect()

tot
```

```{r}
# origin <- census_dataset |>  
#      filter(
#       NOM_ENT == "Puebla",
#       NOM_MUN == "Acateno",
#       NOM_LOC == "Santa Andrea"
#     ) |>
#   collect()
    
tot
    
    # Extract birth data
    birth_local <- as.numeric(tot[1, "PNACENT"])
    birth_another <- as.numeric(tot[1, "PNACOE"])
    
    # Debugging output
    print(paste("Birth Local:", birth_local))
    print(paste("Birth Another:", birth_another))
    
    # Create ratio dataframe
    ratio_df <- tibble(
      Category = c("Local", "Other"),
      Ratio = c(birth_local, birth_another)
    )
    
    # Calculate percentages
    ratio_df$Percentage <- ratio_df$Ratio / sum(ratio_df$Ratio) * 100
    
    # Plot the pie chart
    gg <- ggplot(ratio_df, aes(x = "", y = Ratio, fill = Category)) +
      geom_bar(stat = "identity", width = 1) +
      coord_polar(theta = "y") +
      theme_void() +
      theme(legend.position = "bottom") +
      scale_fill_manual(values = c("#00BFC4", "#F8766D")) +
      geom_text(aes(label = paste0(round(Percentage), "%")), 
                position = position_stack(vjust = 0.5),
                size = 5, color = "white", fontface = "bold")
    
    gg
    
```

### Read CSV from URL

```{r}
test_csv <- read.csv("https://raw.github.ubc.ca/MDS-2023-24/DSCI_532_individual-assignment_marcony1/master/data/processed/data_coords.csv?token=GHSAT0AAAAAAAAACLO5WK36WJ6S74E3NU4CZRJJJGA")

test_csv

```

```{r}
filtered_data <- filter(test_csv, NOM_ENT == "Puebla")
filtered_data
```

### Download with curl

```{r}
library(curl)

url <- "https://raw.github.ubc.ca/MDS-2023-24/DSCI_532_individual-assignment_marcony1/master/data/processed/data_coords.csv?token=GHSAT0AAAAAAAAACLO5WK36WJ6S74E3NU4CZRJJJGA"

response <- curl::curl_fetch_memory(url)

if (response$status_code == 200) {
  csv_content <- rawToChar(response$content)
    test_csv <- read.csv(text = csv_content)
print(test_csv)
}
```

```{r}
library(RCurl)
x <- getURL("https://raw.github.ubc.ca/MDS-2023-24/DSCI_532_individual-assignment_marcony1/master/data/processed/data_coords.csv?token=GHSAT0AAAAAAAAACLO5WK36WJ6S74E3NU4CZRJJJGA")
y <- read.csv(text = x)
y
```

```{r}
# test_2 <- arrow::read_parquet("https://github.ubc.ca/MDS-2023-24/DSCI_532_individual-assignment_marcony1/tree/master/data/processed/parquet_data_coords.parquet")
```
